+++ scripts/gamemodes/supersmashburger.lua
@@ -1,0 +1,8 @@
+return {
+	Name = "supersmashburger",
+	Base = "deathmatch",
+	Description = "Super Smash Burger, be the last burger alive",
+	Properties = {
+		{ Name = "minplayercount", Type = PropertyType.Integer, Default = 2 },
+	}
+}