+++ scripts/gamemodes/supersmashburger/entities/camera_zone.lua
@@ -1,0 +1,7 @@
+RegisterClientScript()
+
+local entity = ScriptedEntity({
+	Base = "entity_camera_rect",
+	IsNetworked = true,
+	Properties = {}
+})