+++ scripts/gamemodes/supersmashburger/entities/ssb_burger.lua
@@ -1,0 +1,8 @@
+RegisterClientScript()
+
+local entity = ScriptedEntity({
+	Base = "entity_burger",
+	IsNetworked = true,
+	MaxHealth = 1,
+	Properties = {}
+})