+++ scripts/gamemodes/supersmashburger/shared.lua
@@ -1,0 +1,7 @@
+RegisterClientScript("cl_init.lua")
+RegisterClientScript("cl_pcts.lua")
+RegisterClientScript("shared.lua")
+
+local gamemode = ScriptedGamemode()
+
+print("Bonjour !")